refactor: split patches into modules
